--- src/main/java/efub/back/jupjup/domain/review/domain/Review.java
@@ -1,38 +1,0 @@
-package efub.back.jupjup.domain.review.domain;
-
-import efub.back.jupjup.domain.member.domain.Member;
-import efub.back.jupjup.global.BaseTimeEntity;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class Review extends BaseTimeEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "review_id", updatable = false)
-    private Long id;
-
-    @Enumerated(EnumType.STRING)
-    private Badge badge;
-
-    @ManyToOne
-    private Member member;
-
-    @Builder
-    public Review(Badge badge, Member member) {
-        this.badge = badge;
-        this.member = member;
-    }
-}
