--- src/main/java/efub/back/jupjup/domain/review/dto/response/Top3BadgeResDto.java
@@ -1,13 +1,0 @@
-package efub.back.jupjup.domain.review.dto.response;
-
-import efub.back.jupjup.domain.review.domain.Badge;
-import java.util.List;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class Top3BadgeResDto {
-    private Integer badgeCount;
-    private List<BadgeInfo> badges;
-}
