+++ src/main/java/efub/back/jupjup/domain/score/exception/BadgeNotExistsForCodeException.java
@@ -1,4 +1,4 @@
-package efub.back.jupjup.domain.review.exception;
+package efub.back.jupjup.domain.score.exception;
 
 import efub.back.jupjup.global.exception.custom.BadRequestException;
 import lombok.NoArgsConstructor;
