+++ src/main/java/efub/back/jupjup/domain/score/exception/NotValidScoreException.java
@@ -1,0 +1,8 @@
+package efub.back.jupjup.domain.score.exception;
+
+import efub.back.jupjup.global.exception.custom.BadRequestException;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+public class NotValidScoreException extends BadRequestException {
+}
