+++ src/main/java/efub/back/jupjup/domain/score/exception/ScoringNotAllowedException.java
@@ -1,8 +1,8 @@
-package efub.back.jupjup.domain.review.exception;
+package efub.back.jupjup.domain.score.exception;
 
 import efub.back.jupjup.global.exception.custom.BadRequestException;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
-public class ReviewNotAllowedException extends BadRequestException {
+public class ScoringNotAllowedException extends BadRequestException {
 }
