+++ src/main/java/efub/back/jupjup/global/exception/ExceptionType.java
@@ -20,8 +20,8 @@ import efub.back.jupjup.domain.post.exception.MismatchPostGenderException;
 import efub.back.jupjup.domain.post.exception.PostNotFoundException;
 import efub.back.jupjup.domain.post.exception.PostjoinNotFoundException;
 import efub.back.jupjup.domain.post.exception.WrongImageFormatException;
-import efub.back.jupjup.domain.review.exception.BadgeNotExistsForCodeException;
-import efub.back.jupjup.domain.review.exception.ReviewNotAllowedException;
+import efub.back.jupjup.domain.score.exception.NotValidScoreException;
+import efub.back.jupjup.domain.score.exception.ScoringNotAllowedException;
 import efub.back.jupjup.domain.security.exception.BlockedAccountException;
 import efub.back.jupjup.domain.security.exception.ExpiredTokenException;
 import efub.back.jupjup.domain.trashCan.exception.FeedbackNotExistsForCodeException;
@@ -49,9 +49,9 @@ public enum ExceptionType {
 	ALREADY_LOGOUT_EXCEPTION("C2001", "이미 로그아웃된 회원입니다.", AlreadyLogoutException.class),
 	EXPIRED_TOKEN_EXCEPTION("C2002", "만료된 토큰입니다.", ExpiredTokenException.class),
 
-	// review 관련 - C3***
-	BADGE_NOT_EXISTS_FOR_CODE_EXCEPTION("C3000", "해당 코드와 일치하는 뱃지가 존재하지 않습니다.", BadgeNotExistsForCodeException.class),
-	REVIEW_NOT_ALLOWED_EXCEPTION("C3001", "참여 멤버가 아니면 리뷰를 작성할 수 없습니다.", ReviewNotAllowedException.class),
+	// score 관련 - C3***
+	SCORING_NOT_ALLOWED_EXCEPTION("C3001", "참여 멤버가 아니면 리뷰를 작성할 수 없습니다.", ScoringNotAllowedException.class),
+	NOT_VALID_SCORE_EXCEPTION("C3002", "플로깅 평점은 1~5 사이의 정수값만 가능합니다.", NotValidScoreException.class),
 
 	// trashCan 관련 - C4***
 	FEEDBACK_NOT_EXISTS_FOR_CODE_EXCEPTION("C4000", "해당 코드와 일치하는 쓰레기통 피드백이 존재하지 않습니다.",
